+++ Cargo.toml
@@ -3,6 +3,7 @@
 members = [
     "mitm_proxy",
     "proxyapi",
+    "proxyapi_models",
     "tauri-ui",
     "tauri-ui/src-tauri",
 ]
