+++ proxyapi_models/src/lib.rs
@@ -1,0 +1,149 @@
+use bytes::Bytes;
+use http::{HeaderMap, Method, StatusCode, Uri, Version};
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ProxiedRequest {
+    #[serde(with = "http_serde::method")]
+    method: Method,
+    #[serde(with = "http_serde::uri")]
+    uri: Uri,
+    #[serde(with = "http_serde::version")]
+    version: Version,
+    #[serde(with = "http_serde::header_map")]
+    headers: HeaderMap,
+    body: Bytes,
+    time: i64,
+}
+
+impl ProxiedRequest {
+    pub fn new(
+        method: Method,
+        uri: Uri,
+        version: Version,
+        headers: HeaderMap,
+        body: Bytes,
+        time: i64,
+    ) -> Self {
+        Self {
+            method,
+            uri,
+            version,
+            headers,
+            body,
+            time,
+        }
+    }
+
+    pub fn method(&self) -> &Method {
+        &self.method
+    }
+
+    pub fn uri(&self) -> &Uri {
+        &self.uri
+    }
+
+    pub fn version(&self) -> &Version {
+        &self.version
+    }
+
+    pub fn headers(&self) -> &HeaderMap {
+        &self.headers
+    }
+
+    pub fn body(&self) -> &Bytes {
+        &self.body
+    }
+
+    pub fn time(&self) -> i64 {
+        self.time
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ProxiedResponse {
+    #[serde(with = "http_serde::status_code")]
+    status: StatusCode,
+    #[serde(with = "http_serde::version")]
+    version: Version,
+    #[serde(with = "http_serde::header_map")]
+    headers: HeaderMap,
+    body: Bytes,
+    time: i64,
+}
+
+impl ProxiedResponse {
+    pub fn new(
+        status: StatusCode,
+        version: Version,
+        headers: HeaderMap,
+        body: Bytes,
+        time: i64,
+    ) -> Self {
+        Self {
+            status,
+            version,
+            headers,
+            body,
+            time,
+        }
+    }
+
+    pub fn status(&self) -> &StatusCode {
+        &self.status
+    }
+
+    pub fn version(&self) -> &Version {
+        &self.version
+    }
+
+    pub fn headers(&self) -> &HeaderMap {
+        &self.headers
+    }
+
+    pub fn body(&self) -> &Bytes {
+        &self.body
+    }
+
+    pub fn time(&self) -> i64 {
+        self.time
+    }
+}
+
+trait ToString {
+    fn to_string(&self) -> String;
+}
+
+trait ToHashString {
+    fn to_hash_string(&self) -> HashMap<String, String>;
+}
+
+impl ToHashString for HeaderMap {
+    fn to_hash_string(&self) -> HashMap<String, String> {
+        let mut headers: HashMap<String, String> = HashMap::new();
+
+        for (k, v) in self.iter() {
+            headers
+                .insert(k.as_str().to_string(), v.to_str().unwrap().to_string())
+                .unwrap_or("NO header".to_string());
+        }
+        headers
+    }
+}
+
+impl ToString for Version {
+    fn to_string(&self) -> String {
+        match *self {
+            Version::HTTP_09 => "HTTP_09".to_string(),
+            Version::HTTP_10 => "HTTP_10".to_string(),
+            Version::HTTP_11 => "HTTP_11".to_string(),
+            Version::HTTP_2 => "HTTP_2".to_string(),
+            Version::HTTP_3 => "HTTP_3".to_string(),
+            _ => "__NonExhaustive".to_string(),
+        }
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct RequestInfo(pub Option<ProxiedRequest>, pub Option<ProxiedResponse>);
