+++ tauri-ui/Cargo.toml
@@ -12,3 +12,4 @@ gloo-utils = { version = "0.1", features = ["serde"]}
 gloo-timers = "0.2"
 wasm-bindgen-futures = "0.4"
 js-sys = "0.3"
+proxyapi_models = {path = "../proxyapi_models"}
