+++ tauri-ui/src/app.rs
@@ -1,5 +1,6 @@
 use gloo_timers::callback::Timeout;
 use serde::{Deserialize, Serialize};
+use proxyapi_models::RequestInfo;
 use std::net::SocketAddr;
 use wasm_bindgen::prelude::*;
 use wasm_bindgen_futures::spawn_local;
@@ -18,43 +19,45 @@ struct Start {
 
 #[function_component(App)]
 pub fn app() -> Html {
+    let trigger = use_force_update();
     let proxy_state = use_state(|| false);
-    let request_count = use_state(|| 0);
+    let requests = use_mut_ref(Vec::<RequestInfo>::new);
     let onclick = {
         let proxy_state = proxy_state.clone();
-        let requests_count = request_count.clone();
+        let requests = requests.clone();
         {
-            let request_count = request_count.clone();
+            let requests = requests.clone();
             let proxy_state = proxy_state.clone();
             Timeout::new(1_000, move || {
                 spawn_local(async move {
-                    let mut count = *request_count;
                     if *proxy_state {
-                       let value= invoke("plugin:proxy|fetch_request", JsValue::NULL).await;
-                       if !value.is_falsy() {
-                           count +=1;
-                       }
-                       request_count.set(count);
+                        let value = invoke("plugin:proxy|fetch_request", JsValue::NULL).await;
+                        if let Ok(request_info) = value.into_serde() {
+                            let mut r = requests.borrow_mut();
+                            r.push(request_info);
+                        }
+                        trigger.force_update();
                     }
-                }
-                )
+                })
             })
             .forget();
         };
 
         Callback::from(move |_| {
+            let requests = requests.clone();
             if *proxy_state {
                 spawn_local(async move {
-                   invoke("plugin:proxy|stop_proxy", JsValue::NULL).await;
+                    invoke("plugin:proxy|stop_proxy", JsValue::NULL).await;
                 });
-                requests_count.set(0);
+                let mut r = requests.borrow_mut();
+                r.drain(..);
             } else {
                 let args = JsValue::from_serde(&Start {
                     addr: "127.0.0.1:8100".parse().unwrap(),
                 })
                 .unwrap();
                 spawn_local(async move {
-                   invoke("plugin:proxy|start_proxy", args).await;
+                    invoke("plugin:proxy|start_proxy", args).await;
                 });
             };
             proxy_state.set(!*proxy_state);
@@ -65,7 +68,7 @@ pub fn app() -> Html {
             <h1>{"Man In The Middle Proxy"}</h1>
             <button {onclick}>{if *proxy_state {"On"} else {"Off"} }</button>
             if *proxy_state {
-                <p>{*request_count}</p>
+                <p>{format!("{:?}",*requests)}</p>
             }
         </main>
     }
